(WIP) Add home & rules
This commit lays the groundwork for adding the home and rules
text views to the app. The rules are unfinished and I need to
figure out how to swap between them on menu selection and how
to display the appropriate context menu at the bottom for
text navigation.

Figuring this out should lay the groundwork for the rest of the
application.

Co-authored-by: Chris Gardner <chris@chrislgardner.dev>